refactor(modcache): manually resolve `GOMODCACHE` for Wasm environment
Since `exec.Command` is not available in Wasm environment, we resolve
`GOMODCACHE` by directly checking environment variables following Go's
standard behavior.
